# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: HRPay API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /payroll/runs:
    get:
      operationId: listPayrollRuns
      tags: [payroll]
      summary: List payroll runs
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, processing, completed, failed]
      responses:
        "200":
          description: List of payroll runs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PayrollRun"
    post:
      operationId: createPayrollRun
      tags: [payroll]
      summary: Create a payroll run
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePayrollRunBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayrollRun"
  /payroll/runs/{id}:
    get:
      operationId: getPayrollRun
      tags: [payroll]
      summary: Get a payroll run
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payroll run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayrollRun"
  /payroll/runs/{id}/process:
    post:
      operationId: processPayrollRun
      tags: [payroll]
      summary: Process a payroll run
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayrollRun"
  /payroll/stubs:
    get:
      operationId: listPayStubs
      tags: [payroll]
      summary: List pay stubs
      parameters:
        - name: employeeId
          in: query
          schema:
            type: integer
        - name: payrollRunId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of pay stubs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PayStub"

  # --- TIME TRACKING ---
  /time/entries:
    get:
      operationId: listTimeEntries
      tags: [timeTracking]
      summary: List time entries
      parameters:
        - name: employeeId
          in: query
          schema:
            type: integer
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected]
      responses:
        "200":
          description: List of time entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimeEntry"
    post:
      operationId: createTimeEntry
      tags: [timeTracking]
      summary: Create a time entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTimeEntryBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeEntry"
  /time/entries/{id}/approve:
    post:
      operationId: approveTimeEntry
      tags: [timeTracking]
      summary: Approve a time entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeEntry"

  # --- LEAVE ---
...
  /onboarding/tasks:
    get:
      operationId: listOnboardingTasks
      tags: [onboarding]
      summary: List onboarding tasks for an employee
      parameters:
        - name: employeeId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Onboarding tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OnboardingTask"
    post:
      operationId: createOnboardingTask
      tags: [onboarding]
      summary: Create an onboarding task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOnboardingTaskBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnboardingTask"
  /onboarding/tasks/{id}/complete:
    post:
      operationId: completeOnboardingTask
      tags: [onboarding]
      summary: Mark an onboarding task as complete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnboardingTask"
components:
  schemas:
...
    PayrollRun:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        periodStart:
          type: string
          format: date
        periodEnd:
          type: string
          format: date
        payDate:
          type: string
          format: date
        status:
          type: string
          enum: [draft, processing, completed, failed]
        totalGrossPay:
          type: number
        totalNetPay:
          type: number
        totalTaxes:
          type: number
        totalDeductions:
          type: number
        employeeCount:
          type: integer
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
        processedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - periodStart
        - periodEnd
        - payDate
        - status
    CreatePayrollRunBody:
      type: object
      properties:
        name:
          type: string
        periodStart:
          type: string
          format: date
        periodEnd:
          type: string
          format: date
        payDate:
          type: string
          format: date
        notes:
          type: string
      required:
        - name
        - periodStart
        - periodEnd
        - payDate
    PayStub:
      type: object
      properties:
        id:
          type: integer
        payrollRunId:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: string
        grossPay:
          type: number
        netPay:
          type: number
        federalTax:
          type: number
        stateTax:
          type: number
        socialSecurity:
          type: number
        medicare:
          type: number
        healthInsurance:
          type: number
        retirement:
          type: number
        hoursWorked:
          type: number
        periodStart:
          type: string
          format: date
        periodEnd:
          type: string
          format: date
        payDate:
          type: string
          format: date
      required:
        - id
        - payrollRunId
        - employeeId
        - grossPay
        - netPay
    TimeEntry:
      type: object
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: string
        date:
          type: string
          format: date
        clockIn:
          type: string
          format: date-time
        clockOut:
          type: string
          format: date-time
        hoursWorked:
          type: number
        overtimeHours:
          type: number
        notes:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - employeeId
        - date
        - status
    CreateTimeEntryBody:
      type: object
      properties:
        employeeId:
          type: integer
        date:
          type: string
          format: date
        clockIn:
          type: string
          format: date-time
        clockOut:
          type: string
          format: date-time
        notes:
          type: string
      required:
        - employeeId
        - date
        - clockIn
...
    OnboardingTask:
      type: object
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: string
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [documentation, it_setup, training, introduction, compliance, benefits]
        dueDate:
          type: string
          format: date
        completedAt:
          type: string
          format: date-time
        isCompleted:
          type: boolean
        assignedTo:
          type: string
        priority:
          type: string
          enum: [low, medium, high]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - employeeId
        - title
        - category
        - isCompleted
    CreateOnboardingTaskBody:
      type: object
      properties:
        employeeId:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [documentation, it_setup, training, introduction, compliance, benefits]
        dueDate:
          type: string
          format: date
        assignedTo:
          type: string
        priority:
          type: string
          enum: [low, medium, high]
      required:
        - employeeId
        - title
        - category