# lib/api-zod/src/generated/types/payStub.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HRPay API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PayStub {
  id: number;
  payrollRunId: number;
  employeeId: number;
  employeeName?: string;
  grossPay: number;
  netPay: number;
  federalTax?: number;
  stateTax?: number;
  socialSecurity?: number;
  medicare?: number;
  healthInsurance?: number;
  retirement?: number;
  hoursWorked?: number;
  periodStart?: Date;
  periodEnd?: Date;
  payDate?: Date;
}